Complete Admin Login

+++ SERVER/src/database/migrations/_migrations.js
@@ -1,37 +1,32 @@
 import conn from "#config/db.config.js";
 import color, { catchErr } from "#color";
-
-const table = process.argv[2];
-const command = process.argv[3];
-const rbTable = ["images", "events", "users"];
-function consoleColor() {
-    color(
-        ["Command Line Argument ERROR", "red", ["underline", "bold"]],
-        ["\nRun Following Command Example", "yellow", "italic"],
-        ["\n\n-> npm run migration <table_name> <command>"],
-        ["\n\n<commandlist>", "bold"],
-        [
-`
-create 
-drop
-drop_create
-seeders
-addcolumn "<column_name> <data_type>"
-destroyAll`,
-        ]
-    );
-}
-
+import { getFiles, consoleColor } from "#untils/migrationFunctions.js";
+import e from "express";
+const command = process.argv[2];
+const table = process.argv[3];
+const fileList = await getFiles();
 if (!table || !command) {
     consoleColor();
     process.exit(0);
 }
-
+var i = null;
 async function loadModule(t, c) {
+    await fileList.find((value, index, array) => {
+        if (value[0] == t) {
+            i = index;
+        }
+    });
     try {
-        const module = t !== "all" ? await import(`./${t}.migration.js`) : null;
+        var module;
+        try {
+            module =
+                (t !== "all" && c !== "sql")
+                    ?  await import(`./${fileList[i][0]}_${fileList[i][1]}.js`)
+                    : null;
+        } catch (error) {
+            throw "Not Found FIle -> " + t;
+        }
         let query = "";
-
         switch (c) {
             case "create":
                 query = module.createTable;
@@ -50,16 +45,21 @@ async function loadModule(t, c) {
                 query = module.destroyAll;
                 break;
             case "rollback":
-                await rollback();
+                await rollback(t);
                 return;
             case "addcolumn":
                 query = `${module.AddColumn} ${process.argv[4]}`;
+                break;
+            case "sql":
+                conn.query(t,(err,result)=>{
+                    if(err) return console.error(err.sqlMessage);
+                    console.log("Result:",result);
+                });
                 break;
             default:
                 consoleColor();
                 process.exit(0);
         }
-
         if (query) {
             conn.query(query, (err) => {
                 if (err) return console.error(err.sqlMessage);
@@ -71,44 +71,77 @@ async function loadModule(t, c) {
         process.exit(0);
     }
 }
-
-async function rollback() {
-    for (const tbl of rbTable) {
+async function rollback(t) {
+    if (t != "all") {
         try {
-            const module = await import(`./${tbl}.migration.js`);
+            const module = await import(
+                `./${fileList[i][0]}_${fileList[i][1]}.js`
+            );
             conn.query(module.dropTable, (err) => {
                 if (err) return console.log("SQL ERROR => ", err.sqlMessage);
-                color(["Sucssessfuly Drop  " + tbl, "red", "bold"]);
+                color(["Sucssessfuly Drop  " + fileList[i][0], "red", "bold"]);
             });
-        } catch (err) {
-            catchErr("_migration/rollback", err);
-        }
-    }
-
-    for (const tbl of rbTable.slice().reverse()) {
-        try {
-            const module = await import(`./${tbl}.migration.js`);
             conn.query(module.createTable, (err) => {
                 if (err) return console.log("SQL ERROR => ", err.sqlMessage);
-                color(["--Successfully created " + tbl, "green", "bold"]);
+                color([
+                    "--Successfully created " + fileList[i][0],
+                    "green",
+                    "bold",
+                ]);
+            });
+            conn.query(module.seeders, (err) => {
+                if (err) return console.log("SQL ERROR => ", err.sqlMessage);
+                color([
+                    "--Successfully run Seeders " + fileList[i][0],
+                    "yellow",
+                    "bold",
+                ]);
+            });
+        } catch (error) {
+            console.log("ERROR", error);
+        }
+        return;
+    }
+
+    for (const tbl of fileList.slice().reverse()) {
+        try {
+            const module = await import(`./${tbl[0]}_${tbl[1]}.js`);
+            conn.query(module.dropTable, (err) => {
+                if (err) return console.log("SQL ERROR => ", err.sqlMessage);
+                color(["Sucssessfuly Drop  " + tbl[0], "red", "bold"]);
             });
         } catch (err) {
             catchErr("_migration/rollback", err);
         }
     }
 
-    for (const tbl of rbTable.slice().reverse()) {
+    for (const tbl of fileList) {
         try {
-            const module = await import(`./${tbl}.migration.js`);
+            const module = await import(`./${tbl[0]}_${tbl[1]}.js`);
+            conn.query(module.createTable, (err) => {
+                if (err) return console.log("SQL ERROR => ", err.sqlMessage);
+                color(["--Successfully created " + tbl[0], "green", "bold"]);
+            });
+        } catch (err) {
+            catchErr("_migration/rollback", err);
+        }
+    }
+
+    for (const tbl of fileList) {
+        try {
+            const module = await import(`./${tbl[0]}_${tbl[1]}.js`);
             conn.query(module.seeders, (err) => {
                 if (err) return console.log("SQL ERROR => ", err.sqlMessage);
-                color(["--Successfully run Seeders " + tbl, "yellow", "bold"]);
+                color([
+                    "--Successfully run Seeders " + tbl[0],
+                    "yellow",
+                    "bold",
+                ]);
             });
         } catch (err) {
             catchErr("_migration/rollback", err);
         }
     }
 }
-
 await loadModule(table, command);
 conn.end();
